Workprogress en Home Modificado Logo de navbar

+++ frontend/src/app/services/auth.service.ts
@@ -1,5 +1,6 @@
+import { isPlatformBrowser } from '@angular/common';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { Injectable } from '@angular/core';
+import { Inject, Injectable, PLATFORM_ID } from '@angular/core';
 import { User } from '../models/user.model';
 
 @Injectable({
@@ -9,7 +10,11 @@ export class AuthService {
 
   private url:string = 'http://localhost:3000';
 
-  constructor(private http:HttpClient) {
+  private isBrowser:boolean = false;
+  isLoggedIn:boolean = false;//Variable para que pueda ser usada por AppComponent y habilitar el navbar
+
+  constructor(private http:HttpClient, @Inject(PLATFORM_ID) platformId: Object) {
+    this.isBrowser = isPlatformBrowser(platformId);
   }
 
   async signIn(user: User){
@@ -39,8 +44,10 @@ export class AuthService {
   }
 
   signOut(){
-    localStorage.clear();
-    window.location.reload();
+    if(this.isBrowser){
+      localStorage.clear();
+      window.location.reload();
+    }
   }
 
 
@@ -48,20 +55,36 @@ export class AuthService {
     if(token != null && token != undefined && token != ''){
       const headerType = 'x-access-token';
       return await this.http.get(`${this.url}/api/auth/checkToken`, { headers: {[headerType]: token} }).toPromise().then( res => {
+        this.isLoggedIn = true;
         return res;
       }).catch(error => {
+        this.isLoggedIn = false;
         return error;
       });
     }else{
+      this.isLoggedIn = false;
       return false;
     }
-    
   }
+
   saveToken(token: string){
-    localStorage.setItem('signIn', token);
+    if(this.isBrowser){
+      localStorage.setItem('signIn', token);
+    }
   }
   getToken(){
-    const token = localStorage.getItem('signIn');
-    return token;
+    if(this.isBrowser){
+      const token = localStorage.getItem('signIn');
+      return token;
+    }
+  }
+
+
+  async getPosts(){
+    const headerType = 'x-access-token';
+    const token = this.getToken();
+    return await this.http.get(`${this.url}/api/posts`, { headers: {[headerType]: token} }).toPromise().then( res => {
+      return res;
+    });
   }
 }
